project/requirements.txt
Pillow==8.4.0
rembg==1.7.0